New plugin system, file struct, python3 compat
