+++ .gitignore
@@ -8,3 +8,4 @@ test/
 /dist/
 /src/*.egg-info/
 /build/
+/**/__pycache__/
