+++ server/src/rspet/server/CLI_state_diagram.png
